+++ user_service/requirements.txt
@@ -1,0 +1,3 @@
+grpcio
+grpcio-tools
+Flask
